Main

Description

Este modulo recebe o tabuleiro de entrada que através da várias funçoes separa o seus elementos e os trabalha até ao final de todos os movimentos do robot e retorna o rusultado final

Documentation

type Board = (Mapa, Pos, Ori, Movs, Save, Tick, Lights, NLights, Size)

Declaração da variavel Board

type Mapa = [String]

Declaração do array de String que contem o Mapa

type Pos = (Int, Int)

Declação da variavel Pos, Tuplo que vai contar uma posição do Mapa

type Ori = Char

Declaração da variavel Ori que indica a orientaçao do robot

type Movs = String

Declaração da variavel Movs que contem a string dos movimentos

type Size = Pos

Declaração da variavel Size, guarda a maior posiçao do tabuleiro

type NLights = Int

Declaração da variavel NLights, que vai conter o numero de luzes do tabuleiro

type Lights = [Pos]

Declaração variavel Lights, contar uma arrays de Pos com todas as posiçoes das luzes

type Tick = Int

Declaração da variavel Tick, conta o numero de movimentos validos

type Save = String

Declaração da varievel Save, String usada para salvar a posiçao do robot sempre que acender ou apagar uma luz

main :: IO ()

Função que recebe do Input e devolve o Output do programa

geraBoard :: [String] -> String

Esta função é responvel por passar um array de Strings para uma variavel do tipo Board que é usada para chamar a função mov

getBot :: [String] -> String

Função que procura no Array a String que contem a posição e orietação do robot e devolve essa String

getBoard :: [String] -> Mapa

Função que retira todas as String anterioes a String que contem um espaço (string da posiçao do robot) e devolve esse array de strings (Mapa)

getPos :: String -> Pos

Função que recebe uma string em que os dois primeiros elementos sao a posição do robot e devolve esses mesmos elementos num tuplo

getOri :: String -> Ori

Recebe uma String com a posiçao e orientaçao do robot separados por espaços e devolve a orientaçao do robot que é a ultima posiçao da string

getLights

Arguments

:: Mapa
Mapa contento as luzes e niveis do caminho

-> Int
Inteiro usadao para saber qual a cordenada y

-> Lights
Array de luzes

Recebe uma Mapa e chama uma funçao auxiliar que devolve um array de luzes

getLights'

Arguments

:: String
String pertencente ao Mapa

-> (Int, Int)
Cordenadas

-> Lights
Array de luzes

Função auxiliar da funçao getLights que percorre a string a procura de luzes e devolve a posição das luzes num array

mov :: Board -> String

Função recursiva responsavel pelo ciclo que vai percorrer toda a String de movimentos, chamar a função responsavel por casa movimento e devolve o resultado final

jumpCheck

Arguments

:: Char
Nivel actual

-> Char
Nivel da Proxima posição

-> Bool
Devolve verdadeiro oui falto

Verifica se é possivel da o salto

Propriedades do Salto:

So pode saltar se a proxima posição for 1 nivel acima;

Nunca pode saltar para o mesmo nivel;

Podes saltar para qualquer nivel desde que seja inferior ao nivel inicial ou seja pode saltar do nivel z para o a.

rodaD :: Ori -> Ori

Função para fazer uma rotação a direita ( Recebe a oricientaçao actual devolve a nova)

rodaE :: Ori -> Ori

Função para fazer uma rotação a esquerda ( Recebe a oricientaçao actual devolve a nova)

findPos :: Mapa -> Pos -> Size -> Char

getLinha :: Mapa -> Pos -> (Int, Int) -> Char

getCol :: String -> Pos -> Int -> Char

isLight :: Pos -> Lights -> Bool

turnLight :: Mapa -> Pos -> Int -> Mapa

turnLight' :: String -> Int -> Int -> String